(#55) YellowAnimaBehaviorをテンプレートに移動
