--- Assets/Scripts/YellowAnimaMovementBehavior.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfca7282015103444ad5abb2ef3e760f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
